impl From<MultiSignedEthEvent> for EthMsgUpdate
